Avaliações/Trabalhos/Trabalho3-front-react/api/config.py
import os
from dotenv import load_dotenv

load_dotenv()
basedir = os.path.abspath(os.path.dirname(__file__))
class Config: #configs do banco e orm
    SQLALCHEMY_DATABASE_URI = f"sqlite:///{os.path.join(basedir, 'database.db')}"
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    SECRET_KEY = os.getenv('SECRET_KEY', 'chave ultra secreta')
    JWT_SECRET_KEY = os.getenv('SECRET_KEY', 'chave ultra secreta')
