Avaliações/Trabalhos/Trabalho3-front-react/api/requirements.txt
api==0.0.7
Flask==3.1.0
Flask_Cors==5.0.0
Flask_JWT_Extended==4.6.0
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
